php script is added on the nav bar that test the website if user is still logged in
on the php script, the user will have to log in first. if the user directly enter the index.php or the other page, the user will be redirected to the log in page. and if the user is still logged in, a hyperlink text will be displayed "You are still logged in"when the user clicks, the user will be redirected on the recent page you open